+++ neurocaps/_utils/_logger.py
@@ -1,5 +1,6 @@
 """Internal logging function and class for flushing"""
 import logging, sys
+from logging.handlers import QueueHandler
 
 # Global variables to determine if a handler is user defined or defined by OS
 _USER_ROOT_HANDLER = None
@@ -10,7 +11,7 @@ class _Flush(logging.StreamHandler):
         super().emit(record)
         self.flush()
 
-def _logger(name, flush=False, top_level=True):
+def _logger(name, flush=False, top_level=True, parallel_log_config=None):
     global _USER_ROOT_HANDLER, _USER_MODULE_HANDLERS
 
     logger = logging.getLogger(name.split(".")[-1])
@@ -22,6 +23,13 @@ def _logger(name, flush=False, top_level=True):
         _USER_ROOT_HANDLER = logging.getLogger().hasHandlers()
         _USER_MODULE_HANDLERS[logger.name] = logging.getLogger(logger.name).hasHandlers()
 
+    # Special case for parallel config to pass a user-defined logger specifically for parallel processing.
+    if logger.name == "_extract_timeseries" and not top_level and parallel_log_config:
+        # Only QueueHandler will be in handler list
+        logger.handlers.clear()
+        if "queue" in parallel_log_config: logger.addHandler(QueueHandler(parallel_log_config["queue"]))
+        if "level" in parallel_log_config: logger.setLevel(parallel_log_config["level"])
+
     if not logger.level: logger.setLevel(logging.INFO)
 
     # Check if user defined root handler or assigned a specific handler for module
@@ -31,7 +39,7 @@ def _logger(name, flush=False, top_level=True):
     logger.propagate = True if _USER_ROOT_HANDLER else False
 
     # Add or messages will repeat several times due to multiple handlers if same name used
-    if not default_handlers and not (logger.name == "_extract_timeseries" and top_level):
+    if not default_handlers and not (parallel_log_config or logger.name == "_extract_timeseries" and top_level):
         # If no user specified default handler, any handler is assigned by OS and is cleared
         if logger.name == "_extract_timeseries": logger.handlers.clear()